+++ x/stakers/types/codec.go
@@ -1,6 +1,7 @@
 package types
 
 import (
+	"github.com/KYVENetwork/chain/x/stakers/types_v1beta1"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codecTypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cryptoCodec "github.com/cosmos/cosmos-sdk/crypto/codec"
@@ -21,6 +22,7 @@ func RegisterInterfaces(registry codecTypes.InterfaceRegistry) {
 	registry.RegisterImplementations((*sdk.Msg)(nil), &MsgJoinPool{})
 	registry.RegisterImplementations((*sdk.Msg)(nil), &MsgLeavePool{})
 	registry.RegisterImplementations((*sdk.Msg)(nil), &MsgUpdateParams{})
+	registry.RegisterImplementations((*sdk.Msg)(nil), &types_v1beta1.MsgUpdateParams{})
 }
 
 var Amino = codec.NewLegacyAmino()